refactor files to new structure
