+++ devfiles/src/ui/OptionsComp/GlobalOptionsComp.tsx
@@ -1,5 +1,5 @@
-import '../App.css';
-import PageManager from '../classes/PageManager.js';
+import '../../App.css';
+import PageManager from '../../classes/PageManager.js';
 
 function GlobalOptionsComp(params: { pageManager: PageManager }) {
     return <>Global Options</>;
